Added new associative array for second field of /etc/passwd
This introduces a new variable default:parsed_etc_passwd_shadow._passwd_passwd
as an associative string array mapping each user to the second field in /etc/passwd.

This is useful for developing inventory about the use of /etc/shadow for each
local user.

The variable is underscore prefixed and explicitly tagged for noreport to
prevent it from being sent back to an Enterprise Hub.